replace payment portal redirect
