+++ frontend/src/pages/org/index.ts
@@ -38,7 +38,6 @@ import "./browser-profiles-list";
 import "./browser-profiles-new";
 import "./settings/settings";
 import "./dashboard";
-import "./payment-portal-redirect";
 
 const RESOURCE_NAMES = ["workflow", "collection", "browser-profile", "upload"];
 type ResourceName = (typeof RESOURCE_NAMES)[number];
@@ -77,7 +76,6 @@ export type OrgParams = {
   settings: {
     settingsTab?: "information" | "members";
   };
-  "payment-portal-redirect": {};
 };
 export type OrgTab = keyof OrgParams;
 
@@ -310,9 +308,6 @@ export class Org extends LiteElement {
         }
         // falls through
       }
-      case "payment-portal-redirect":
-        tabPanelContent = this.renderPaymentPortalRedirect();
-        break;
       default:
         tabPanelContent = html`<btrix-not-found
           class="flex items-center justify-center"
@@ -724,14 +719,6 @@ export class Org extends LiteElement {
       @org-remove-member=${this.onOrgRemoveMember}
       @btrix-update-org=${this.updateOrg}
     ></btrix-org-settings>`;
-  }
-
-  private renderPaymentPortalRedirect() {
-    return html`<btrix-org-payment-portal-redirect
-      class="flex flex-1"
-      orgId=${this.orgId}
-      .authState=${this.authState}
-    ></btrix-org-payment-portal-redirect>`;
   }
 
   private async onSelectNewDialog(e: SelectNewDialogEvent) {
